🎉 feat(karium.config.ts): add build configuration file for Karium 🚀 chore(package.json): add build and release scripts, add files field 🔧 chore(tsconfig.json): change path for freshdeps
The Karium build configuration file is added to the project to define the build process. The build script is added to the package.json file to build the project. The release script is added to the package.json file to test, generate changelog, publish the package, and push the tags to the repository. The files field is added to the package.json file to specify the files that should be included in the package. The path for freshdeps is changed in the tsconfig.json file to match the new name of the binary.